[FLINK-7959] [table] Split CodeGenerator into CodeGeneratorContext and ExprCodeGenerator.
