+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/codegen/FunctionCodeGenerator.scala
@@ -21,75 +21,49 @@ import org.apache.flink.api.common.functions._
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.configuration.Configuration
 import org.apache.flink.streaming.api.functions.ProcessFunction
-import org.apache.flink.table.api.TableConfig
 import org.apache.flink.table.codegen.CodeGenUtils.{boxedTypeTermForTypeInfo, newName}
 import org.apache.flink.table.codegen.Indenter.toISC
 
 /**
   * A code generator for generating Flink [[org.apache.flink.api.common.functions.Function]]s.
-  * Including [[MapFunction]], [[FlatMapFunction]], [[FlatJoinFunction]], [[ProcessFunction]].
-  *
-  * @param config configuration that determines runtime behavior
-  * @param nullableInput input(s) can be null.
-  * @param input1 type information about the first input of the Function
-  * @param input2 type information about the second input if the Function is binary
-  * @param input1FieldMapping additional mapping information for input1
-  *   (e.g. POJO types have no deterministic field order and some input fields might not be read)
-  * @param input2FieldMapping additional mapping information for input2
-  *   (e.g. POJO types have no deterministic field order and some input fields might not be read)
+  * Including [[MapFunction]], [[FlatMapFunction]], [[FlatJoinFunction]], [[ProcessFunction]], and
+  * the corresponding rich version of the functions.
   */
-class FunctionCodeGenerator(
-    config: TableConfig,
-    nullableInput: Boolean,
-    input1: TypeInformation[_ <: Any],
-    input2: Option[TypeInformation[_ <: Any]] = None,
-    input1FieldMapping: Option[Array[Int]] = None,
-    input2FieldMapping: Option[Array[Int]] = None)
-  extends CodeGenerator(
-    config,
-    nullableInput,
-    input1,
-    input2,
-    input1FieldMapping,
-    input2FieldMapping) {
-
-  /**
-    * A code generator for generating unary Flink
-    * [[org.apache.flink.api.common.functions.Function]]s with one input.
-    *
-    * @param config configuration that determines runtime behavior
-    * @param nullableInput input(s) can be null.
-    * @param input type information about the input of the Function
-    * @param inputFieldMapping additional mapping information necessary for input
-    *   (e.g. POJO types have no deterministic field order and some input fields might not be read)
-    */
-  def this(
-    config: TableConfig,
-    nullableInput: Boolean,
-    input: TypeInformation[Any],
-    inputFieldMapping: Array[Int]) =
-    this(config, nullableInput, input, None, Some(inputFieldMapping))
+object FunctionCodeGenerator {
 
   /**
     * Generates a [[org.apache.flink.api.common.functions.Function]] that can be passed to Java
     * compiler.
     *
+    * @param ctx The context of the code generator
     * @param name Class name of the Function. Must not be unique but has to be a valid Java class
     *             identifier.
     * @param clazz Flink Function to be generated.
     * @param bodyCode code contents of the SAM (Single Abstract Method). Inputs, collector, or
     *                 output record can be accessed via the given term methods.
     * @param returnType expected return type
+    * @param input1Type the first input type
+    * @param input1Term the first input term
+    * @param input2Type the second input type, optional.
+    * @param input2Term the second input term.
+    * @param collectorTerm the collector term
+    * @param contextTerm the context term
     * @tparam F Flink Function to be generated.
     * @tparam T Return type of the Flink Function.
     * @return instance of GeneratedFunction
     */
   def generateFunction[F <: Function, T <: Any](
-    name: String,
-    clazz: Class[F],
-    bodyCode: String,
-    returnType: TypeInformation[T])
-  : GeneratedFunction[F, T] = {
+      ctx: CodeGeneratorContext,
+      name: String,
+      clazz: Class[F],
+      bodyCode: String,
+      returnType: TypeInformation[T],
+      input1Type: TypeInformation[_ <: Any],
+      input1Term: String = CodeGeneratorContext.DEFAULT_INPUT1_TERM,
+      input2Type: Option[TypeInformation[_ <: Any]] = None,
+      input2Term: Option[String] = Some(CodeGeneratorContext.DEFAULT_INPUT2_TERM),
+      collectorTerm: String = CodeGeneratorContext.DEFAULT_COLLECTOR_TERM,
+      contextTerm: String = CodeGeneratorContext.DEFAULT_CONTEXT_TERM): GeneratedFunction[F, T] = {
     val funcName = newName(name)
 
     // Janino does not support generics, that's why we need
@@ -98,7 +72,7 @@ class FunctionCodeGenerator(
     // FlatMapFunction
     if (clazz == classOf[FlatMapFunction[_, _]]) {
       val baseClass = classOf[RichFlatMapFunction[_, _]]
-      val inputTypeTerm = boxedTypeTermForTypeInfo(input1)
+      val inputTypeTerm = boxedTypeTermForTypeInfo(input1Type)
       (baseClass,
         s"void flatMap(Object _in1, org.apache.flink.util.Collector $collectorTerm)",
         List(s"$inputTypeTerm $input1Term = ($inputTypeTerm) _in1;"))
@@ -107,7 +81,7 @@ class FunctionCodeGenerator(
     // MapFunction
     else if (clazz == classOf[MapFunction[_, _]]) {
       val baseClass = classOf[RichMapFunction[_, _]]
-      val inputTypeTerm = boxedTypeTermForTypeInfo(input1)
+      val inputTypeTerm = boxedTypeTermForTypeInfo(input1Type)
       (baseClass,
         "Object map(Object _in1)",
         List(s"$inputTypeTerm $input1Term = ($inputTypeTerm) _in1;"))
@@ -116,31 +90,31 @@ class FunctionCodeGenerator(
     // FlatJoinFunction
     else if (clazz == classOf[FlatJoinFunction[_, _, _]]) {
       val baseClass = classOf[RichFlatJoinFunction[_, _, _]]
-      val inputTypeTerm1 = boxedTypeTermForTypeInfo(input1)
-      val inputTypeTerm2 = boxedTypeTermForTypeInfo(input2.getOrElse(
+      val inputTypeTerm1 = boxedTypeTermForTypeInfo(input1Type)
+      val inputTypeTerm2 = boxedTypeTermForTypeInfo(input2Type.getOrElse(
         throw new CodeGenException("Input 2 for FlatJoinFunction should not be null")))
       (baseClass,
         s"void join(Object _in1, Object _in2, org.apache.flink.util.Collector $collectorTerm)",
         List(s"$inputTypeTerm1 $input1Term = ($inputTypeTerm1) _in1;",
-             s"$inputTypeTerm2 $input2Term = ($inputTypeTerm2) _in2;"))
+             s"$inputTypeTerm2 ${input2Term.get} = ($inputTypeTerm2) _in2;"))
     }
 
     // JoinFunction
     else if (clazz == classOf[JoinFunction[_, _, _]]) {
       val baseClass = classOf[RichJoinFunction[_, _, _]]
-      val inputTypeTerm1 = boxedTypeTermForTypeInfo(input1)
-      val inputTypeTerm2 = boxedTypeTermForTypeInfo(input2.getOrElse(
+      val inputTypeTerm1 = boxedTypeTermForTypeInfo(input1Type)
+      val inputTypeTerm2 = boxedTypeTermForTypeInfo(input2Type.getOrElse(
         throw new CodeGenException("Input 2 for JoinFunction should not be null")))
       (baseClass,
         s"Object join(Object _in1, Object _in2)",
         List(s"$inputTypeTerm1 $input1Term = ($inputTypeTerm1) _in1;",
-          s"$inputTypeTerm2 $input2Term = ($inputTypeTerm2) _in2;"))
+          s"$inputTypeTerm2 ${input2Term.get} = ($inputTypeTerm2) _in2;"))
     }
 
     // ProcessFunction
     else if (clazz == classOf[ProcessFunction[_, _]]) {
       val baseClass = classOf[ProcessFunction[_, _]]
-      val inputTypeTerm = boxedTypeTermForTypeInfo(input1)
+      val inputTypeTerm = boxedTypeTermForTypeInfo(input1Type)
       (baseClass,
         s"void processElement(Object _in1, " +
           s"org.apache.flink.streaming.api.functions.ProcessFunction.Context $contextTerm," +
@@ -156,30 +130,30 @@ class FunctionCodeGenerator(
       public class $funcName
           extends ${samHeader._1.getCanonicalName} {
 
-        ${reuseMemberCode()}
+        ${ctx.reuseMemberCode()}
 
         public $funcName() throws Exception {
-          ${reuseInitCode()}
+          ${ctx.reuseInitCode()}
         }
 
-        ${reuseConstructorCode(funcName)}
+        ${ctx.reuseConstructorCode(funcName)}
 
         @Override
         public void open(${classOf[Configuration].getCanonicalName} parameters) throws Exception {
-          ${reuseOpenCode()}
+          ${ctx.reuseOpenCode()}
         }
 
         @Override
         public ${samHeader._2} throws Exception {
           ${samHeader._3.mkString("\n")}
-          ${reusePerRecordCode()}
-          ${reuseInputUnboxingCode()}
+          ${ctx.reusePerRecordCode()}
+          ${ctx.reuseInputUnboxingCode()}
           $bodyCode
         }
 
         @Override
         public void close() throws Exception {
-          ${reuseCloseCode()}
+          ${ctx.reuseCloseCode()}
         }
       }
     """.stripMargin
