+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/codegen/calls/FunctionGenerator.scala
@@ -569,7 +569,7 @@ object FunctionGenerator {
       method: Method)
     : Unit = {
     sqlFunctions((sqlOperator, operandTypes)) =
-      new NotCallGenerator(new MethodCallGen(BOOLEAN_TYPE_INFO, method))
+      new NotCallGen(new MethodCallGen(BOOLEAN_TYPE_INFO, method))
   }
 
   private def addSqlFunction(
