+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/codegen/calls/RandCallGen.scala
@@ -20,7 +20,7 @@ package org.apache.flink.table.codegen.calls
 
 import org.apache.flink.api.common.typeinfo.BasicTypeInfo._
 import org.apache.flink.table.codegen.calls.CallGenerator.generateCallIfArgsNotNull
-import org.apache.flink.table.codegen.{CodeGenerator, GeneratedExpression}
+import org.apache.flink.table.codegen.{CodeGeneratorContext, GeneratedExpression}
 
 /**
   * Generates a random function call.
@@ -29,26 +29,25 @@ import org.apache.flink.table.codegen.{CodeGenerator, GeneratedExpression}
 class RandCallGen(isRandInteger: Boolean, hasSeed: Boolean) extends CallGenerator {
 
   override def generate(
-      codeGenerator: CodeGenerator,
-      operands: Seq[GeneratedExpression])
-    : GeneratedExpression = {
-
+      ctx: CodeGeneratorContext,
+      operands: Seq[GeneratedExpression],
+      nullCheck: Boolean): GeneratedExpression = {
     val randField = if (hasSeed) {
       if (operands.head.literal) {
-        codeGenerator.addReusableRandom(Some(operands.head))
+        ctx.addReusableRandom(Some(operands.head), nullCheck)
       } else {
         s"(new java.util.Random(${operands.head.resultTerm}))"
       }
     } else {
-      codeGenerator.addReusableRandom(None)
+      ctx.addReusableRandom(None, nullCheck)
     }
 
     if (isRandInteger) {
-      generateCallIfArgsNotNull(codeGenerator.nullCheck, INT_TYPE_INFO, operands) { terms =>
+      generateCallIfArgsNotNull(nullCheck, INT_TYPE_INFO, operands) { terms =>
         s"$randField.nextInt(${terms.last})"
       }
     } else {
-      generateCallIfArgsNotNull(codeGenerator.nullCheck, DOUBLE_TYPE_INFO, operands) { _ =>
+      generateCallIfArgsNotNull(nullCheck, DOUBLE_TYPE_INFO, operands) { _ =>
         s"$randField.nextDouble()"
       }
     }
